show surrounding shapes in tldraw annotation view
